test16

+++ run_django_linux.py
@@ -1,6 +1,7 @@
 import subprocess
 import os
 import sys
+import time
 import psycopg2
 from psycopg2 import sql
 from dotenv import load_dotenv
@@ -17,6 +18,7 @@ def run_command(command):
 
 def recreate_postgres_db():
     try:
+        # Подключаемся к базе postgres
         conn = psycopg2.connect(
             dbname='postgres',  
             user=os.getenv("LOCAL_DB_USER"),
@@ -26,8 +28,35 @@ def recreate_postgres_db():
         )
         conn.autocommit = True
         cursor = conn.cursor()
+
+        # Завершаем активные подключения к базе данных, которую нужно удалить
+        print("Завершаем активные подключения к базе данных...")
+        cursor.execute(sql.SQL("""
+            SELECT pg_terminate_backend(pg_stat_activity.pid)
+            FROM pg_stat_activity
+            WHERE pg_stat_activity.datname = %s
+            AND pid <> pg_backend_pid();
+        """), [os.getenv("LOCAL_DB_NAME")])
+
+        # Проверяем, что все процессы завершены
+        time.sleep(2)  # Даем немного времени для завершения процессов
+        cursor.execute("""
+            SELECT count(*) FROM pg_stat_activity WHERE datname = %s;
+        """, [os.getenv("LOCAL_DB_NAME")])
+        result = cursor.fetchone()
+        
+        if result[0] > 0:
+            print(f"Не удалось завершить все подключения к базе данных: {result[0]} активных подключений.")
+            sys.exit(1)
+
+        # Удаление базы данных
+        print("Удаляем базу данных...")
         cursor.execute(sql.SQL("DROP DATABASE IF EXISTS {}").format(sql.Identifier(os.getenv("LOCAL_DB_NAME"))))
+
+        # Создание новой базы данных
+        print("Создаем базу данных заново...")
         cursor.execute(sql.SQL("CREATE DATABASE {}").format(sql.Identifier(os.getenv("LOCAL_DB_NAME"))))
+
         cursor.close()
         conn.close()
         print(f"База данных {os.getenv('LOCAL_DB_NAME')} успешно пересоздана.")
@@ -35,9 +64,10 @@ def recreate_postgres_db():
         print(f"Ошибка при перезапуске базы данных PostgreSQL: {e}")
         sys.exit(1)
 
-# Перезапуск
+# Перезапуск базы данных
 recreate_postgres_db()
 
+# Выполнение миграций и других команд
 run_command('python3 delete_migrations.py')
 run_command('python3 manage.py makemigrations')
 run_command('python3 manage.py migrate')
@@ -62,5 +92,5 @@ run_command('python3 manage.py shell -c "exec(open(\'create_superuser.py\').read
 os.remove('create_superuser.py')
 print('Суперпользователь успешно создан.')
 
-# runserver
-subprocess.run('python3 manage.py runserver', shell=True)
+# Запуск сервера
+subprocess.run('python3 manage.py runserver & python3 manage.py startbot', shell=True)
